# #158 refactor : optimize orm queries in place views