… places/views.py