place_search : 239쿼리 118개의 중복 쿼리 454.18ms -> 121쿼리 118개의 중복 쿼리 145.0ms
place_detail : 9쿼리 3개의 중복 쿼리 9.61ms -> 6쿼리 15.28ms